requirements.txt
nltk~=3.7
torch~=1.11.0
torchvision~=0.12.0
torchaudio~=0.11.0
pika~=1.2.1
transformers~=4.19.2
pydantic~=1.9.1
fastapi~=0.78.0
uvicorn~=0.17.6
pyyaml==6.0